move to sub projects
more refactoring

refactor a bit

add unit tests
