+++ core/src/main/scala/com/broadinstitute/dsp/package.scala
@@ -1,0 +1,8 @@
+package com.broadinstitute
+
+import org.broadinstitute.dsde.workbench.google2.{RegionName, ZoneName}
+
+package object dsp {
+  val zoneName = ZoneName("us-central1-a")
+  val regionName = RegionName("us-central1")
+}
