+++ core/src/test/resources/reference.conf
@@ -1,0 +1,9 @@
+path-to-credential = "path-to-credential"
+
+database {
+  url = "jdbc:mysql://127.0.0.1:3306/leonardo?rewriteBatchedStatements=true&nullNamePatternMatchesAll=true"
+  user = "username"
+  password = "password"
+}
+
+report-destination-bucket = "qi-test"
