+++ core/src/test/scala/com/broadinstitute/dsp/ConfigSpec.scala
@@ -1,0 +1,24 @@
+package com.broadinstitute.dsp
+
+import java.nio.file.Paths
+
+import org.broadinstitute.dsde.workbench.model.google.GcsBucketName
+import org.scalatest.flatspec.AnyFlatSpec
+import org.scalatest.matchers.should.Matchers
+
+class ConfigSpec extends AnyFlatSpec with Matchers {
+  it should "read config file correctly" in {
+    val config = Config.appConfig
+    val expectedConfig = AppConfig(
+      DatabaseConfig(
+        "jdbc:mysql://127.0.0.1:3306/leonardo?rewriteBatchedStatements=true&nullNamePatternMatchesAll=true",
+        "username",
+        "password"
+      ),
+      Paths.get("path-to-credential"),
+      GcsBucketName("qi-test")
+    )
+
+    config shouldBe Right(expectedConfig)
+  }
+}
