+++ resource-validator/src/main/resources/META-INF/native-image/com/broadinstitute/dsp/resourcevalidator/reflect-config.json
