+++ resource-validator/src/main/resources/logback.xml
