+++ resource-validator/src/main/resources/reference.conf
