+++ resource-validator/src/test/resources/reference.conf
