--- src/main/scala/com/broadinstitute/dsp/resourcevalidator/package.scala
@@ -1,8 +1,0 @@
-package com.broadinstitute.dsp
-
-import org.broadinstitute.dsde.workbench.google2.{RegionName, ZoneName}
-
-package object resourcevalidator {
-  val zoneName = ZoneName("us-central1-a")
-  val regionName = RegionName("us-central1")
-}
