+++ zombie-monitor/src/main/resources/reference.conf
@@ -1,0 +1,9 @@
+path-to-credential = ${LEONARDO_PATH_TO_CREDENTIAL}
+
+database {
+  url = "jdbc:mysql://127.0.0.1:3306/leonardo?rewriteBatchedStatements=true&nullNamePatternMatchesAll=true"
+  user = ${LEONARDO_DB_USER}
+  password = ${LEONARDO_DB_PASSWORD}
+}
+
+report-destination-bucket = ${REPORT_DESTINATION_BUCKET}
