+++ zombie-monitor/src/test/resources/reference.conf
@@ -1,0 +1,8 @@
+path-to-credential = "path-to-credential"
+
+database {
+  user = "username"
+  password = "password"
+}
+
+report-destination-bucket = "qi-test"
